chore(clients/go): move utils into internal
