+++ clients/go/internal/utils/jsonString_mixin.go
