+++ clients/go/internal/utils/structmap.go
