+++ clients/go/internal/utils/structmap_test.go
