+++ clients/go/internal/utils/test_utils.go
