+++ clients/go/pkg/commands/resolveIncident_command.go
@@ -20,7 +20,7 @@ import (
 	"time"
 
 	"github.com/zeebe-io/zeebe/clients/go/internal/pb"
-	"github.com/zeebe-io/zeebe/clients/go/utils"
+	"github.com/zeebe-io/zeebe/clients/go/internal/utils"
 )
 
 type DispatchResolveIncidentCommand interface {
