+++ clients/go/pkg/commands/resolveIncident_command_test.go
@@ -21,7 +21,7 @@ import (
 	"github.com/golang/mock/gomock"
 	"github.com/zeebe-io/zeebe/clients/go/internal/mock_pb"
 	"github.com/zeebe-io/zeebe/clients/go/internal/pb"
-	"github.com/zeebe-io/zeebe/clients/go/utils"
+	"github.com/zeebe-io/zeebe/clients/go/internal/utils"
 )
 
 func TestResolveIncidentCommand(t *testing.T) {
