+++ clients/go/pkg/worker/jobDispatcher_test.go
@@ -17,9 +17,9 @@ package worker
 import (
 	"github.com/stretchr/testify/suite"
 	"github.com/zeebe-io/zeebe/clients/go/internal/pb"
+	"github.com/zeebe-io/zeebe/clients/go/internal/utils"
 	"github.com/zeebe-io/zeebe/clients/go/pkg/commands"
 	"github.com/zeebe-io/zeebe/clients/go/pkg/entities"
-	"github.com/zeebe-io/zeebe/clients/go/utils"
 	"sync"
 	"testing"
 	"time"
