+++ clients/go/pkg/worker/jobWorker_test.go
@@ -21,9 +21,9 @@ import (
 	"github.com/golang/protobuf/proto"
 	"github.com/zeebe-io/zeebe/clients/go/internal/mock_pb"
 	"github.com/zeebe-io/zeebe/clients/go/internal/pb"
+	"github.com/zeebe-io/zeebe/clients/go/internal/utils"
 	"github.com/zeebe-io/zeebe/clients/go/pkg/commands"
 	"github.com/zeebe-io/zeebe/clients/go/pkg/entities"
-	"github.com/zeebe-io/zeebe/clients/go/utils"
 	"testing"
 	"time"
 )
